+++ src/main/java/ru/mels/tasktracking/exception/EmployeeNotFoundException.java
@@ -1,12 +1,7 @@
 package ru.mels.tasktracking.exception;
 
-/**
- * @author Meleshkin Alexandr
- * @since 07.03.2022
- */
-
 public class EmployeeNotFoundException extends NotFoundElementException {
-    public EmployeeNotFoundException(Integer id) {
+    public EmployeeNotFoundException(Long id) {
         super("Employee with id - " + id + " is not found");
     }
 }
