+++ src/main/java/ru/mels/tasktracking/exception/ProjectNotFoundException.java
@@ -1,12 +1,8 @@
 package ru.mels.tasktracking.exception;
 
-/**
- * @author Meleshkin Alexandr
- * @since 07.03.2022
- */
 public class ProjectNotFoundException extends NotFoundElementException{
 
-    public ProjectNotFoundException(Integer id) {
+    public ProjectNotFoundException(Long id) {
         super("Project with id - " + id + " is not found");
     }
 }
