+++ src/main/java/ru/mels/tasktracking/exception/TaskBoardNotFoundException.java
@@ -1,11 +1,7 @@
 package ru.mels.tasktracking.exception;
 
-/**
- * @author Meleshkin Alexandr
- * @since 07.03.2022
- */
 public class TaskBoardNotFoundException extends NotFoundElementException{
-    public TaskBoardNotFoundException(Integer id) {
+    public TaskBoardNotFoundException(Long id) {
         super("Task board with id - " + id + " is not found");
     }
 }
