+++ src/main/java/ru/mels/tasktracking/exception/TaskNotFoundException.java
@@ -1,11 +1,7 @@
 package ru.mels.tasktracking.exception;
 
-/**
- * @author Meleshkin Alexandr
- * @since 07.03.2022
- */
 public class TaskNotFoundException extends NotFoundElementException{
-    public TaskNotFoundException(Integer id) {
+    public TaskNotFoundException(Long id) {
         super("Task with id - " + id + " is not found");
     }
 }
