+++ src/main/java/ru/mels/tasktracking/repository/ProjectRepository.java
@@ -3,9 +3,5 @@ package ru.mels.tasktracking.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import ru.mels.tasktracking.entity.Project;
 
-/**
- * @author Meleshkin Alexandr
- * @since 07.03.2022
- */
-public interface ProjectRepository extends JpaRepository<Project, Integer> {
+public interface ProjectRepository extends JpaRepository<Project, Long> {
 }
